// src/main/webapp/src/app/rest/model/githubUserActivityStatsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GithubUserActivityMonthStatsDTO } from './githubUserActivityMonthStatsDTO';


export interface GithubUserActivityStatsDTO { 
    user?: string;
    githubPullRequestCreatedCount?: number;
    githubPullRequestUpdatedCount?: number;
    githubPullRequestCommentedCount?: number;
    githubPullRequestMergedCount?: number;
    githubPullRequestClosedCount?: number;
    perMonth?: { [key: string]: GithubUserActivityMonthStatsDTO; };
}