// src/main/webapp/src/app/rest/model/jiraUserActivityStatsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JiraUserActivityMonthStatsDTO } from './jiraUserActivityMonthStatsDTO';


export interface JiraUserActivityStatsDTO { 
    user?: string;
    jiraIssueCreatedCount?: number;
    jiraIssueUpdatedCount?: number;
    jiraIssueCommentedCount?: number;
    jiraIssueClosedRejectedCount?: number;
    jiraIssueCloseResolvedCount?: number;
    perMonth?: { [key: string]: JiraUserActivityMonthStatsDTO; };
}