// src/main/webapp/src/app/rest/model/mailMessageUserActivityStatsDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MailMessageUserActivityMonthStatsDTO } from './mailMessageUserActivityMonthStatsDTO';


export interface MailMessageUserActivityStatsDTO { 
    user?: string;
    mailMessageSentCount?: number;
    mailMessageRepliedCount?: number;
    mailMessageVotedCount?: number;
    perMonth?: { [key: string]: MailMessageUserActivityMonthStatsDTO; };
}